feat #201: 상점 조회 여부 구현
- internal 구조로 수정
- internal 상점 조회 여부 구현
