+++ shop-service/src/main/java/com/musinsam/shopservice/presentation/controller/ShopInternalControllerApiV1.java
@@ -4,7 +4,7 @@ import com.musinsam.shopservice.application.dto.request.ReqShopGetSearchDtoApiV1
 import com.musinsam.shopservice.application.dto.response.ResInternalShopGetByShopIdDtoApiV1;
 import com.musinsam.shopservice.application.dto.response.ResShopCouponDtoApiV1;
 import com.musinsam.shopservice.application.dto.response.ResShopGetDtoApiV1;
-import com.musinsam.shopservice.application.service.ShopServiceApiV1;
+import com.musinsam.shopservice.application.service.ShopInternalServiceApiV1;
 import jakarta.validation.Valid;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
@@ -24,7 +24,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/internal/v1/shops")
 public class ShopInternalControllerApiV1 {
 
-  private final ShopServiceApiV1 shopServiceApiV1;
+  private final ShopInternalServiceApiV1 shopInternalServiceApiV1;
 
   /**
    * 상점 단건 조회
@@ -33,7 +33,7 @@ public class ShopInternalControllerApiV1 {
   public ResponseEntity<ResInternalShopGetByShopIdDtoApiV1> internalGetShop(
       @PathVariable UUID shopId
   ) {
-    return ResponseEntity.ok(shopServiceApiV1.internalGetShop(shopId));
+    return ResponseEntity.ok(shopInternalServiceApiV1.internalGetShop(shopId));
   }
 
   /**
@@ -44,7 +44,7 @@ public class ShopInternalControllerApiV1 {
       @Valid Pageable pageable,
       @RequestBody ReqShopGetSearchDtoApiV1.Shop dto
   ) {
-    return ResponseEntity.ok(shopServiceApiV1.internalGetShopListSearch(pageable, dto));
+    return ResponseEntity.ok(shopInternalServiceApiV1.internalGetShopListSearch(pageable, dto));
   }
 
   /**
@@ -54,6 +54,17 @@ public class ShopInternalControllerApiV1 {
   public ResponseEntity<ResShopCouponDtoApiV1> internalCouponGetByShopId(
       @RequestParam UUID shopId
   ) {
-    return ResponseEntity.ok(shopServiceApiV1.internalCouponGetByShopId(shopId));
+    return ResponseEntity.ok(shopInternalServiceApiV1.internalCouponGetByShopId(shopId));
   }
+
+  /**
+   * 상점 존재 여부 조회
+   */
+  @GetMapping("/{shopId}/exists")
+  public ResponseEntity<Boolean> internalExistsShopById(
+      @PathVariable UUID shopId
+  ) {
+    return ResponseEntity.ok(shopInternalServiceApiV1.internalExistsShopById(shopId));
+  }
+
 }
